+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG:
+* 0.10.12
+    * Add instrument to collect data from jemalloc. Requires feature `jemalloc-ctl`.
+    * Impl `From<Snapshot>` for `ItemKind.
 * 0.10.11
     * Transfer the meter from metrics 0.2.3 here since the crate in that repository has changed to a whole new project [commit](https://github.com/chridou/metrix/commit/693e86e839b8870bcbfae93f1416d094ee2e88a6)
 * 0.10.10